+++ chipnomad_lib/project.c
@@ -11,7 +11,7 @@ FXName fxNamesCommon[] = {
   {fxARP, "ARP"}, {fxARC, "ARC"}, {fxPVB, "PVB"}, {fxPBN, "PBN"}, {fxPSL, "PSL"}, {fxPIT, "PIT"}, {fxPRD, "PRD"},
   {fxVOL, "VOL"}, {fxRET, "RET"}, {fxDEL, "DEL"}, {fxOFF, "OFF"}, {fxKIL, "KIL"},
   {fxTIC, "TIC"}, {fxTBL, "TBL"}, {fxTBX, "TBX"}, {fxTHO, "THO"}, {fxTXH, "TXH"},
-  {fxGRV, "GRV"}, {fxGGR, "GGR"}, {fxHOP, "HOP"},
+  {fxGRV, "GRV"}, {fxGGR, "GGR"}, {fxHOP, "HOP"}, {fxSNG, "SNG"}
 };
 int fxCommonCount = sizeof(fxNamesCommon) / sizeof(FXName);